// src/api/model/getListingsParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GetListingsPropertyType } from './getListingsPropertyType';
import type { Pageable } from './pageable';

export type GetListingsParams = {
  /**
   * Filter by property type
   */
  propertyType?: GetListingsPropertyType;
  /**
   * Minimum price (inclusive)
   */
  priceMin?: number;
  /**
   * Maximum price (inclusive)
   */
  priceMax?: number;
  /**
   * Filter by district name
   */
  district?: string;
  /**
   * Minimum number of bedrooms
   */
  minBedrooms?: number;
  pageable: Pageable;
};